separated config loading to file
