+++ dspace/config/vsb/fetch-vocabularies.py
@@ -15,37 +15,9 @@ from pathlib import Path
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from threading import Lock
 import configparser
+from vsb_config import VSB_BASE_URL, VSB_TEST_URL
 
 # Configuration
-
-def load_vsb_config():
-    """Return vsb.base.url and vsb.test.url from config/dspace.cfg."""
-    # --- Locate the config file ---
-    current = Path(__file__).resolve().parent
-    while current != current.parent:
-        cfg = current / "config" / "dspace.cfg"
-        if cfg.exists():
-            break
-        current = current.parent
-    else:
-        raise FileNotFoundError("Could not find config/dspace.cfg")
-
-    # --- Read the two URLs ---
-    vsb_base, vsb_test = None, None
-    with cfg.open(encoding="utf-8") as f:
-        for line in f:
-            if line.strip().startswith("vsb.base.url"):
-                vsb_base = line.split("=", 1)[1].strip()
-            elif line.strip().startswith("vsb.test.url"):
-                vsb_test = line.split("=", 1)[1].strip()
-
-    if not (vsb_base and vsb_test):
-        raise ValueError("Missing vsb.base.url or vsb.test.url in dspace.cfg")
-    else:
-        print(f"Found VSB URLs: {vsb_base}, {vsb_test}")
-    return vsb_base, vsb_test
-
-VSB_BASE_URL, VSB_TEST_URL = load_vsb_config()
 TIMEOUT = 10  # Reduced timeout
 MAX_WORKERS = 8  # Concurrent downloads
 
@@ -267,7 +239,6 @@ Examples:
     successful_downloads = 0
     total_conversions = 0
     successful_conversions = 0
-    return
     # Process vocabularies
     with ThreadPoolExecutor(max_workers=MAX_WORKERS) as executor:
         futures = []
